1.1.0/requirements.txt
docx==0.2.4
jsonpath==0.75
pdfminer3k==1.3.1
Pillow==5.2.0
PyExecJS==1.5.1
pymongo==3.6.1
PyMySQL==0.8.1
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyquery==1.4.0
requests==2.18.4
Scrapy==1.5.0
PyYAML==3.13
PyMuPDF==1.13.20
redis==2.10.6
Js2Py==0.59
demjson==2.2.4
redis_py_cluster==1.3.6
fonttools==3.41.0
pyppeteer==0.0.25
fonttools==3.41.0
selenium==3.141.0

